Made the entirety of rat city aka the rat town square
I created the town square in a separate scene. added a lot of lighting to create the mood and vibe. Also added a few new props. For example the noodle cups, the box houses, the can restaurant, drink box houses and the pringles apartment. The church building is also set at place. And besides that, there's now a little centerpiece in the townsquare.