created the environment for the dumpster
created the environment for the dumpster in a separate dumpster scene. I also deleted a wire prefab because it was broken. Hopefully that doesn't cause any issue. also the light halo effect on top of the trash heap is only visible when there's a camera in the scene apparently. I mean I think it's a nice touch. Let me know if any further adjustments needs to be done